+++ coinbase/rest/common.py
@@ -5,7 +5,7 @@ from coinbase.constants import API_PREFIX
 
 def get_unix_time(self, **kwargs) -> Dict[str, Any]:
     """
-    **Get UNIX Time**
+    **Get Server Time**
     _________________
     [GET] https://api.coinbase.com/api/v3/brokerage/time
 
@@ -13,12 +13,13 @@ def get_unix_time(self, **kwargs) -> Dict[str, Any]:
 
     **Description:**
 
-    Get the current time from the Coinbase Advanced API.
+    Get the current time from the Coinbase Advanced API. This is a public endpoint.
 
     __________
 
-    **Read more on the official documentation:** `Get UNIX Time <https://docs.cloud.coinbase.com/advanced-trade-api/reference/retailbrokerageapi_getunixtime>`_
+    **Read more on the official documentation:** `Get Server Time <https://docs.cloud.coinbase.com/advanced-trade-api/reference/retailbrokerageapi_getservertime>`_
     """
+
     endpoint = f"{API_PREFIX}/time"
 
-    return self.get(endpoint, **kwargs)
+    return self.get(endpoint, public=True, **kwargs)