+++ docs/conf.py
@@ -28,3 +28,5 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 html_theme = "sphinx_rtd_theme"
 html_static_path = ["_static"]
+
+autodoc_member_order = "bysource"