+++ tests/rest/test_rest_base.py
@@ -43,6 +43,31 @@ class RestBaseTest(unittest.TestCase):
 
             self.assertEqual(accounts, expected_response)
 
+    def test_get_public(self):
+        client = RESTClient(api_key=TEST_API_KEY, api_secret=TEST_API_SECRET)
+
+        with Mocker() as m:
+            expected_response = {"iso": "2022-01-01T00:00:00Z", "epoch": 1640995200}
+
+            with Mocker() as m:
+                m.request(
+                    "GET",
+                    "https://api.coinbase.com/api/v3/brokerage/time",
+                    json=expected_response,
+                )
+
+                client.get("/api/v3/brokerage/time", public=True)
+
+                captured_request = m.request_history[0]
+                captured_headers = captured_request.headers
+
+                self.assertNotIn("Authorization", captured_headers)
+                self.assertTrue("User-Agent" in captured_headers)
+                self.assertEqual(
+                    captured_headers["User-Agent"],
+                    "coinbase-advanced-py/" + __version__,
+                )
+
     def test_post(self):
         client = RESTClient(api_key=TEST_API_KEY, api_secret=TEST_API_SECRET)